+++ app/models/mattermost_work_package_post.rb
@@ -5,5 +5,4 @@ class MattermostWorkPackagePost < ApplicationRecord
 
   validates :post_id, presence: true
   validates :channel_id, presence: true
-  validates :work_package_id, uniqueness: true
 end